Fixed yet another typo

+++ graph/src/iters.rs
@@ -567,7 +567,7 @@ impl Graph {
         quantity: usize,
         random_state: u64,
     ) -> impl IndexedParallelIterator<Item = NodeT> + '_ {
-        let number_of_nodes = self.get_nodes_number();
+        let number_of_nodes = self.get_nodes_number() as u64;
         (0..quantity).into_par_iter().map(move |i| unsafe {
             sample_uniform(number_of_nodes, splitmix64(random_state + i as u64)) as NodeT
         })
